WEB-1337
All prices are shown with the cents value in the sign up process.

+++ rfidEJB/src/com/n4systems/model/signuppackage/SignUpPackage.java
@@ -74,8 +74,8 @@ public class SignUpPackage {
 		return signPackageDetails.isPreferred();
 	}
 
-	public int getDefaultPricePerUserPerMonth() {
-		return getDefaultPaymentOption().getPricePerUserPerMonth().intValue();
+	public float getDefaultPricePerUserPerMonth() {
+		return getDefaultPaymentOption().getPricePerUserPerMonth();
 	}
 
 	public Long getContract(PaymentOption paymentOption) {
